+++ examples/main.go
@@ -19,6 +19,7 @@ func main() {
 		main_iterate,
 		main_reduce,
 		main_sort,
+		main_seq2_api,
 	}
 
 	for _, f := range funcs {