+++ examples/sort.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"cmp"
 	"fmt"
 	"math/rand"
 
@@ -12,7 +13,7 @@ func main_sort() {
 	fmt.Println("picking up 5 random numbers from higher to lower")
 	stream.Generate(rand.Uint32).
 		Limit(5).
-		Sorted(order.Inverse(order.Natural[uint32])).
+		Sorted(order.Inverse(cmp.Compare[uint32])).
 		ForEach(func(n uint32) {
 			fmt.Println(n)
 		})