+++ stream/instantiate_test.go
@@ -1,6 +1,7 @@
 package stream
 
 import (
+	"cmp"
 	"math/rand"
 	"testing"
 
@@ -52,7 +53,7 @@ func TestOfMap_SortedByKey(t *testing.T) {
 	months := OfMap(map[int]string{
 		1: "Jan", 2: "Feb", 3: "Mar", 4: "Apr", 5: "May", 6: "Jun",
 		7: "Jul", 8: "Aug", 9: "Sep", 10: "Oct", 11: "Nov", 12: "Dec",
-	}).Sorted(order.ByKey[int, string](order.Natural[int]))
+	}).Sorted(order.ByKey[int, string](cmp.Compare[int]))
 
 	monthNames := Map(months, func(p item.Pair[int, string]) string {
 		return p.Val
@@ -71,7 +72,7 @@ func TestOfMap_SortedByVal(t *testing.T) {
 	months := OfMap(map[int]string{
 		1: "Jan", 2: "Feb", 3: "Mar", 4: "Apr", 5: "May", 6: "Jun",
 		7: "Jul", 8: "Aug", 9: "Sep", 10: "Oct", 11: "Nov", 12: "Dec",
-	}).Sorted(order.ByVal[int, string](order.Natural[string]))
+	}).Sorted(order.ByVal[int, string](cmp.Compare[string]))
 
 	monthNames := Map(months, func(p item.Pair[int, string]) string {
 		return p.Val